requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.8.0
mplfinance==0.12.10b0
multidict==6.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
py-cord==2.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
